+++ src/Aardvark.SceneGraph/SgCSharp.fs
@@ -49,14 +49,14 @@ type SceneGraphExtensions =
     static member Surface(sg : ISg, surface : IBackendSurface) = Sg.SurfaceApplicator(Surface.Backend surface, sg) :> ISg
 
     [<Extension>]
-    static member Surface(sg : ISg, effect : FShade.Effect) = Sg.SurfaceApplicator(Surface.FShadeSimple effect, sg) :> ISg
+    static member Surface(sg : ISg, effect : FShade.Effect) = Sg.SurfaceApplicator(Surface.Effect effect, sg) :> ISg
 
     [<Extension>]
-    static member Surface(sg : ISg, [<ParamArray>] effects : FShade.Effect[]) = Sg.SurfaceApplicator(Surface.FShadeSimple (FShade.Effect.compose effects), sg) :> ISg
+    static member Surface(sg : ISg, [<ParamArray>] effects : FShade.Effect[]) = Sg.SurfaceApplicator(Surface.Effect (FShade.Effect.compose effects), sg) :> ISg
 
     [<Extension>]
     static member Surface(sg : ISg, creator : Func<FShade.EffectConfig, FShade.EffectInputLayout*aval<FShade.Imperative.Module>>) =
-        Sg.SurfaceApplicator(Surface.FShade (fun cfg -> creator.Invoke(cfg)), sg) :> ISg
+        Sg.SurfaceApplicator(Surface.Dynamic (fun cfg -> creator.Invoke(cfg)), sg) :> ISg
 
     // Blending
     [<Extension>]